smartgardeningassistant/requirements.txt
Flask
requests
opencv-python
pillow
tensorflow
paho-mqtt
python-dotenv